Move auto-instrumentation back to tracing
