+++ packages/tracing/src/index.ts
@@ -1,4 +1,3 @@
-import { addExtensionMethods } from '@sentry/core';
 export {
   extractTraceparentData,
   getActiveTransaction,
@@ -15,6 +14,7 @@ export {
 } from '@sentry/core';
 export type { SpanStatusType } from '@sentry/core';
 
+import { addExtensionMethods } from './hubextensions';
 import * as Integrations from './integrations';
 
 export type { RequestInstrumentationOptions } from './browser';
